+++ src/TravelRequestWF.Web/Pages/Employee/Detail.cshtml.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace TravelRequestWF.Web.Pages.Employee;
+
+public class DetailModel : PageModel
+{
+    public void OnGet(int id) { }
+}