+++ src/TravelRequestWF.Web/Pages/Employee/Submit.cshtml.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace TravelRequestWF.Web.Pages.Employee;
+
+public class SubmitModel : PageModel
+{
+    public void OnGet() { }
+}