+++ src/TravelRequestWF.Web/Pages/Manager/Review.cshtml.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace TravelRequestWF.Web.Pages.Manager;
+
+public class ReviewModel : PageModel
+{
+    public void OnGet(int id) { }
+}